https://drive.google.com/file/d/1WbmgEuACpLkBCGmgUNYy8IRBq9VtjEbF/view?usp=drive_link
To Supply Leftover Food to Poor

This project focuses on "FoodConnect," an application built on the Salesforce platform, designed to address the challenge of food wastage and hunger. The primary objective is to connect donors with surplus food to organizations or individuals in need, ensuring efficient redistribution of leftover food to the underprivileged. By leveraging Salesforce's robust features, such as cloud-based data management, real-time collaboration, and automated workflows, the application facilitates seamless coordination between food donors, transport providers, and beneficiaries.